+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [Unreleased]
+- Added parameter info support
+
 ## 0.5.0-beta (2026-01-22)
 - Added completion prioritization
 - Added find usages feature