+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added parameter info support
+- Added struct member reference and completion support
 - Minor feature improvements
 
 ## 0.5.0-beta (2026-01-22)