+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Added parameter info support
 - Added struct member reference and completion support
+- Added structure view feature
 - Minor feature improvements
 
 ## 0.5.0-beta (2026-01-22)