+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Added structure view feature
 - Added breadcrumbs support
 - Added structure-aware navigation bar feature
+- Added documentation feature
 - Minor feature improvements
 
 ## 0.5.0-beta (2026-01-22)