+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [Unreleased]
+## 0.6.0-beta (2026-02-07)
 - Added parameter info support
 - Added struct member reference and completion support
 - Added structure view feature